Backend/requirements.txt
pre-commit>=3.3.0
black==23.7.0
isort==5.12.0
flake8==6.0.0
flake8-django==0.1.0